// src/interfaces/professionalOccupation.interface.ts
export interface ProfessionalOccupation {
  availableMinutes: number;
  bookedMinutes: number;
  cancelledAppointments: number;
  completedAppointments: number;
  noShowAppointments: number;
  occupationRate: number;
  professionalId: string;
  professionalName: string;
  totalAppointments: number;
}